Move spritesheet to toolbox
